🐛 fix(server/rooms): 메세지가 없고 마지막 유저가 방을 나갔을 때 터지는 이슈 해결 (#357)
#357

+++ packages/server/src/chats/rooms.service.ts
@@ -516,7 +516,7 @@ export class RoomService {
           room: { id: roomId },
         });
         const ids = msgs.map((msg) => msg.id);
-        await this.messagesRepository.delete(ids);
+        if (ids.length !== 0) await this.messagesRepository.delete(ids);
         await this.roomsRepository.delete(roomId);
         return { isDeleted: true };
       }
